+++ src/log-collector/LogCollectorLaunchInfo.ts
@@ -4,4 +4,5 @@ export interface LogCollectorLaunchInfo {
   shell: string;
   root?: string;
   enableStdin?: boolean;
+  port?: number;
 }